requirements.txt
opencv-contrib-python
PySimpleGUI
Pillow
